Move the viewer probe to `utilities/`, with its browser dependency inline
A debugging tool's dependencies do not belong in the library's. As a PEP 723 script it runs in an
environment `uv` builds for it, so `playwright` reaches neither `pyproject.toml` nor the lockfile,
and nothing in `positronic` can import it.

Ticket: Positronic-Robotics/internal#128 #refs
